batch operation add event, and modify WrapEL
